Move ePUB metadata generation to pre-defined file
The primary motivation is to make as much metadata specification
consistent and not randomly generated by pandoc.

For now this is being used to keep a consistent URN across rebuilds

Some ereaders make use of the URN to dedupe/auto-update books.

In the future it can be used to encode other ePUB metadata in an easy
way